// src/main/models/entity/master/CodebookDetail.ts
import { Sequelize } from 'sequelize';
import { BelongsTo, Column, DataType, ForeignKey, Model, Table } from 'sequelize-typescript';
import Codebook from './Codebook';

export interface CodebookDetailAttributes {
  id: number;
  codebookId: number;
  textCode: string;
  textValue: string;
  textOther: Object;
  description: string;
  indexed: number;
  parentId: number;
  logHistory: Array<Object> | null;
  flag?: number;
  createdBy: string;
  createdAt: Date;
  updatedBy?: string;
  updatedAt?: Date;
  deletedAt?: Date;
  companyId?: string;
  companyInfo?: any;
}

type CodebookDetailAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type CodebookDetailCreationAttributes = Omit<
  CodebookDetailAttributes,
  CodebookDetailAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'codebook_detail',
  underscored: true,
  paranoid: true,
  schema: 'master',
})
class CodebookDetail
  extends Model<CodebookDetailAttributes | CodebookDetailCreationAttributes>
  implements CodebookDetailAttributes {
  @Column({
    allowNull: false,
    type: DataType.BIGINT,
    primaryKey: true,
    autoIncrement: true,
  })
  id!: number;


  @ForeignKey(() => Codebook)
  @Column({
    allowNull: false,
    type: DataType.BIGINT,
  })
  codebookId!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  textCode!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  textValue!: string;

  @Column({
    allowNull: false,
    type: DataType.JSONB,
  })
  textOther!: string;

  @Column({
    allowNull: false,
    type: DataType.TEXT,
  })
  description!: string;

  @Column({
    allowNull: false,
    type: DataType.INTEGER,
  })
  indexed!: number;

  @Column({
    allowNull: false,
    type: DataType.INTEGER,
  })
  parentId!: number;

  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  logHistory!: Array<Object> | null

  @Column({
    type: DataType.INTEGER,
    defaultValue: 1,
  })
  flag!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  updatedBy!: string;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  companyId?: string;

  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  companyInfo?: any;

  @BelongsTo(() => Codebook, {
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  codebook!: Codebook;

}

export default CodebookDetail;